+++ examples/data-sources/meraki_network_vlan_profile_assignment/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_network_vlan_profile_assignment" "example" {
+  network_id = "L_123456"
+}