+++ examples/resources/meraki_network_vlan_profile_assignment/import-by-identity.tf
@@ -1,0 +1,7 @@
+import {
+  to = meraki_network_vlan_profile_assignment.example
+  identity = {
+    "network_id" : "<network_id>"
+    "vlan_profile_iname" : "<vlan_profile_iname>"
+  }
+}