+++ examples/resources/meraki_network_vlan_profile_assignment/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = meraki_network_vlan_profile_assignment.example
+  id = "<network_id>,<vlan_profile_iname>"
+}