+++ examples/resources/meraki_network_vlan_profile_assignment/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_network_vlan_profile_assignment.example "<network_id>,<vlan_profile_iname>"